Revert "fixes"
This reverts commit c821e3ba34cb46d2f5a67c84fcad013192543841.

+++ Source Files/Engine Core/Scene/CUDA/SceneNode.cuh
@@ -10,7 +10,7 @@ public:
 
 	SceneNode(glm::mat4);
 
-	unsigned long bitsUsed;
+	int bitsUsed;
 	uint64 morton;
 
 private:
